For bots and guest: Add a redirect from viewtopic.php?pid= to viewtopic.php?id=
To reduce the endless stream of forum scanning by AI bots (masquerading as users) from subnets of Chinese companies. These bots do not follow the rules in the robots.txt file and crawl all links to posts.
Changed links to the last message in a topic on the topic list and search pages (displayed as a list of topics).
For guests and bots, links to a specific post return a 301 redirect to the topic page where the post is located.